requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pytesseract==0.3.10
qdrant-client==1.9.1
sentence-transformers==3.0.1
redis==5.0.7
python-dotenv==1.0.1
weasyprint==62.3
httpx==0.27.0